% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execInstall.R
\name{execUninstall}
\alias{execUninstall}
\title{Uninstall CLI launchers for package executables}
\usage{
execUninstall(package, destdir = NULL)
}
\arguments{
\item{package}{Package names to process.}

\item{destdir}{Directory containing the launchers. When \code{NULL}, Rapp
determines the installation directory. See \code{\link[Rapp:install_pkg_cli_apps]{Rapp::install_pkg_cli_apps()}} for
information about the defaults directories.}
}
\value{
Invisibly returns the paths of launchers removed by
\code{\link[Rapp:uninstall_pkg_cli_apps]{Rapp::uninstall_pkg_cli_apps()}}.
}
\description{
Remove command-line launchers previously installed for one or more R
packages. This is a convenience wrapper around
\code{\link[Rapp:uninstall_pkg_cli_apps]{Rapp::uninstall_pkg_cli_apps()}}.
}
\seealso{
\code{\link[Rapp:uninstall_pkg_cli_apps]{Rapp::uninstall_pkg_cli_apps()}}, \code{\link[=execInstall]{execInstall()}}
}
