% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execInstall.R
\name{execInstall}
\alias{execInstall}
\title{Install CLI launchers for package executables}
\usage{
execInstall(package, destdir = NULL, lib.loc = NULL, overwrite = NA)
}
\arguments{
\item{package}{Package names to process.}

\item{destdir}{Directory where launchers are written. When \code{NULL}, Rapp
determines the installation directory. See \code{\link[Rapp:install_pkg_cli_apps]{Rapp::install_pkg_cli_apps()}} for
information about the defaults directories.}

\item{lib.loc}{Additional library paths used to locate installed packages.}

\item{overwrite}{Whether existing launchers should be replaced.}
}
\value{
Invisibly returns the paths of launchers written by
\code{\link[Rapp:install_pkg_cli_apps]{Rapp::install_pkg_cli_apps()}}.
}
\description{
Install command-line launchers for the executable scripts provided by one or
more installed R packages. This is a convenience wrapper around
\code{\link[Rapp:install_pkg_cli_apps]{Rapp::install_pkg_cli_apps()}}.
}
\seealso{
\code{\link[Rapp:install_pkg_cli_apps]{Rapp::install_pkg_cli_apps()}}, \code{\link[=execUninstall]{execUninstall()}}
}
