% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execTemplate.R
\name{execTemplate}
\alias{execTemplate}
\title{Copy the CLI template to the scripts directory}
\usage{
execTemplate(scripts_dir, name = "base_template.R")
}
\arguments{
\item{scripts_dir}{Path to the scripts directory.}
}
\value{
Invisibly, the path to the created template script.
}
\description{
Copy the CLI template to the scripts directory
}
